Containerize flask application server
It is important to isolate and breaks the software into pieces
which can be deployed, tested easily and quickly.

Solution:
Docker is the way by which it is possible to deploy software
collectively with its dependencies and its a quicker way. So,
this is possible by using Dockerfile which has set of
instructions of what to install and how to run the software.

Signed-off-by: Madhura Vora <mandemadhura@gmail.com>